Added view.base package and CattleListFragment disassociated from genearl logic
