+++ app/src/main/java/com/iua/agustinpereyra/view/helpviews/HelpActivity.kt
@@ -3,9 +3,8 @@ package com.iua.agustinpereyra.view.helpviews
 import android.content.Intent
 import android.net.Uri
 import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
 import com.iua.agustinpereyra.R
-import com.iua.agustinpereyra.view.BaseActivity
+import com.iua.agustinpereyra.view.base.BaseActivity
 
 class HelpActivity : BaseActivity(), HelpFragment.HelpFragmentListener {
