+++ app/src/main/java/com/iua/agustinpereyra/view/userviews/UserAccountActivity.kt
@@ -3,9 +3,8 @@ package com.iua.agustinpereyra.view.userviews
 import android.app.Activity
 import android.content.Intent
 import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
 import com.iua.agustinpereyra.R
-import com.iua.agustinpereyra.view.BaseActivity
+import com.iua.agustinpereyra.view.base.BaseActivity
 
 class UserAccountActivity : BaseActivity(),
     UserAccountViewFragment.UserAccountFragmentListener,
